Refactor some functions and fixed minor problems
