+++ projeto-kokama/src/api/Api.ts
@@ -1,8 +1,6 @@
 import axios from 'axios';
-import {TRANSLATE_URL} from "../config/urls";
 
 const api = axios.create({
-    baseURL: TRANSLATE_URL,
     timeout: 10*1000,
 });
